+++ packages/front-end/components/Features/RuleModal/RolloutFields.tsx
@@ -1,0 +1,139 @@
+import { useFormContext } from "react-hook-form";
+import { FeatureInterface, FeatureRule } from "back-end/types/feature";
+import { FeatureRevisionInterface } from "back-end/types/feature-revision";
+import Page from "@/components/Modal/Page";
+import Field from "@/components/Forms/Field";
+import FeatureValueField from "@/components/Features/FeatureValueField";
+import RolloutPercentInput from "@/components/Features/RolloutPercentInput";
+import SelectField from "@/components/Forms/SelectField";
+import { NewExperimentRefRule, useAttributeSchema } from "@/services/features";
+import ScheduleInputs from "@/components/Features/ScheduleInputs";
+import { useIncrementer } from "@/hooks/useIncrementer";
+import SavedGroupTargetingField from "@/components/Features/SavedGroupTargetingField";
+import ConditionInput from "@/components/Features/ConditionInput";
+import PrerequisiteTargetingField from "@/components/Features/PrerequisiteTargetingField";
+import {FaExclamationTriangle} from "react-icons/fa";
+
+export default function RolloutFields({
+  feature,
+  environment,
+  defaultValues,
+  version,
+  revisions,
+  setPrerequisiteTargetingSdkIssues,
+  isCyclic,
+  cyclicFeatureId,
+  // legacy:
+  scheduleToggleEnabled,
+  setScheduleToggleEnabled,
+  setShowUpgradeModal,
+}: {
+  feature: FeatureInterface;
+  environment: string;
+  defaultValues: FeatureRule | NewExperimentRefRule;
+  version: number;
+  revisions?: FeatureRevisionInterface[];
+  setPrerequisiteTargetingSdkIssues: (b: boolean) => void;
+  isCyclic: boolean;
+  cyclicFeatureId: string | null;
+
+  scheduleToggleEnabled: boolean;
+  setScheduleToggleEnabled: (b: boolean) => void;
+  setShowUpgradeModal: (b: boolean) => void;
+}) {
+  const form = useFormContext();
+
+  const attributeSchema = useAttributeSchema(false, feature.project);
+  const hasHashAttributes =
+    attributeSchema.filter((x) => x.hashAttribute).length > 0;
+
+  const [conditionKey, forceConditionRender] = useIncrementer();
+
+  return (
+    <>
+      <Page display="Rollout">
+        <Field
+          label="Description"
+          textarea
+          minRows={1}
+          {...form.register("description")}
+          placeholder="Short human-readable description of the rule"
+        />
+
+        <FeatureValueField
+          label="Value to roll out"
+          id="value"
+          value={form.watch("value")}
+          setValue={(v) => form.setValue("value", v)}
+          valueType={feature.valueType}
+          feature={feature}
+          renderJSONInline={true}
+        />
+
+        <div className="appbox mt-4 mb-4 px-3 pt-3 bg-light">
+          <RolloutPercentInput
+            value={form.watch("coverage") || 0}
+            setValue={(coverage) => {
+              form.setValue("coverage", coverage);
+            }}
+            className="mb-1"
+          />
+          <SelectField
+            label="Assign value based on attribute"
+            options={attributeSchema
+              .filter((s) => !hasHashAttributes || s.hashAttribute)
+              .map((s) => ({ label: s.property, value: s.property }))}
+            value={form.watch("hashAttribute")}
+            onChange={(v) => {
+              form.setValue("hashAttribute", v);
+            }}
+            helpText={"The globally unique tracking key for the experiment"}
+          />
+        </div>
+
+        <ScheduleInputs
+          defaultValue={defaultValues.scheduleRules || []}
+          onChange={(value) => form.setValue("scheduleRules", value)}
+          scheduleToggleEnabled={scheduleToggleEnabled}
+          setScheduleToggleEnabled={setScheduleToggleEnabled}
+          setShowUpgradeModal={setShowUpgradeModal}
+          title="Add scheduling to automatically enable/disable this rule"
+        />
+      </Page>
+
+      <Page display="Targeting">
+        <SavedGroupTargetingField
+          value={form.watch("savedGroups") || []}
+          setValue={(savedGroups) => form.setValue("savedGroups", savedGroups)}
+          project={feature.project || ""}
+        />
+        <hr />
+        <ConditionInput
+          defaultValue={form.watch("condition") || ""}
+          onChange={(value) => form.setValue("condition", value)}
+          key={conditionKey}
+          project={feature.project || ""}
+        />
+        <hr />
+        <PrerequisiteTargetingField
+          value={form.watch("prerequisites") || []}
+          setValue={(prerequisites) =>
+            form.setValue("prerequisites", prerequisites)
+          }
+          feature={feature}
+          revisions={revisions}
+          version={version}
+          environments={[environment]}
+          setPrerequisiteTargetingSdkIssues={setPrerequisiteTargetingSdkIssues}
+        />
+        {isCyclic && (
+          <div className="alert alert-danger">
+            <FaExclamationTriangle /> A prerequisite (
+            <code>{cyclicFeatureId}</code>) creates a circular dependency. Remove
+            this prerequisite to continue.
+          </div>
+        )}
+      </Page>
+    </>
+  );
+}
