--- packages/front-end/components/Features/RuleModalNew.tsx
@@ -1,1149 +1,0 @@
-import { useForm } from "react-hook-form";
-import {
-  ExperimentValue,
-  FeatureInterface,
-  FeatureRule,
-  ScheduleRule,
-} from "back-end/types/feature";
-import React, { useMemo, useState } from "react";
-import { date } from "shared/dates";
-import uniqId from "uniqid";
-import {
-  ExperimentInterfaceStringDates,
-  ExperimentType,
-} from "back-end/types/experiment";
-import {
-  getMatchingRules,
-  includeExperimentInPayload,
-  isFeatureCyclic,
-} from "shared/util";
-import { FaExclamationTriangle, FaExternalLinkAlt } from "react-icons/fa";
-import Link from "next/link";
-import cloneDeep from "lodash/cloneDeep";
-import { FeatureRevisionInterface } from "back-end/types/feature-revision";
-import { FaRegCircleCheck } from "react-icons/fa6";
-import { useGrowthBook } from "@growthbook/growthbook-react";
-import {
-  NewExperimentRefRule,
-  generateVariationId,
-  getDefaultRuleValue,
-  getDefaultVariationValue,
-  getFeatureDefaultValue,
-  getRules,
-  useAttributeSchema,
-  useEnvironments,
-  useFeaturesList,
-  validateFeatureRule,
-} from "@/services/features";
-import track from "@/services/track";
-import useOrgSettings from "@/hooks/useOrgSettings";
-import { useExperiments } from "@/hooks/useExperiments";
-import { useDefinitions } from "@/services/DefinitionsContext";
-import { useIncrementer } from "@/hooks/useIncrementer";
-import { useAuth } from "@/services/auth";
-import useSDKConnections from "@/hooks/useSDKConnections";
-import HashVersionSelector, {
-  allConnectionsSupportBucketingV2,
-} from "@/components/Experiment/HashVersionSelector";
-import PrerequisiteTargetingField from "@/components/Features/PrerequisiteTargetingField";
-import Field from "@/components/Forms/Field";
-import Modal from "@/components/Modal";
-import SelectField from "@/components/Forms/SelectField";
-import UpgradeModal from "@/components/Settings/UpgradeModal";
-import StatusIndicator from "@/components/Experiment/StatusIndicator";
-import Toggle from "@/components/Forms/Toggle";
-import { getNewExperimentDatasourceDefaults } from "@/components/Experiment/NewExperimentForm";
-import TargetingInfo from "@/components/Experiment/TabbedPage/TargetingInfo";
-import EditTargetingModal from "@/components/Experiment/EditTargetingModal";
-import ButtonSelectField from "@/components/Forms/ButtonSelectField";
-import PremiumTooltip from "@/components/Marketing/PremiumTooltip";
-import { AppFeatures } from "@/types/app-features";
-import { useUser } from "@/services/UserContext";
-import RadioCards from "@/components/Radix/RadioCards";
-import RadioGroup from "@/components/Radix/RadioGroup";
-import RolloutPercentInput from "./RolloutPercentInput";
-import ConditionInput from "./ConditionInput";
-import FeatureValueField from "./FeatureValueField";
-import NamespaceSelector from "./NamespaceSelector";
-import ScheduleInputs from "./ScheduleInputs";
-import FeatureVariationsInput from "./FeatureVariationsInput";
-import SavedGroupTargetingField from "./SavedGroupTargetingField";
-import FallbackAttributeSelector from "./FallbackAttributeSelector";
-
-export interface Props {
-  close: () => void;
-  feature: FeatureInterface;
-  version: number;
-  setVersion: (version: number) => void;
-  mutate: () => void;
-  i: number;
-  environment: string;
-  defaultType?: string;
-  revisions?: FeatureRevisionInterface[];
-}
-
-export default function RuleModalNew({
-  close,
-  feature,
-  i,
-  mutate,
-  environment,
-  defaultType = "",
-  version,
-  setVersion,
-  revisions,
-}: Props) {
-  const growthbook = useGrowthBook<AppFeatures>();
-  const { hasCommercialFeature } = useUser();
-
-  const attributeSchema = useAttributeSchema(false, feature.project);
-  const [showUpgradeModal, setShowUpgradeModal] = useState(false);
-
-  const { namespaces } = useOrgSettings();
-
-  const rules = getRules(feature, environment);
-  const rule = rules[i];
-
-  const { datasources } = useDefinitions();
-
-  const { experiments, experimentsMap, mutateExperiments } = useExperiments();
-
-  const [allowDuplicateTrackingKey, setAllowDuplicateTrackingKey] = useState(
-    false
-  );
-
-  const [showTargetingModal, setShowTargetingModal] = useState(false);
-
-  const settings = useOrgSettings();
-
-  const defaultRuleValues = getDefaultRuleValue({
-    defaultValue: getFeatureDefaultValue(feature),
-    ruleType: defaultType,
-    attributeSchema,
-  });
-
-  const [conditionKey, forceConditionRender] = useIncrementer();
-
-  const { features } = useFeaturesList();
-  const environments = useEnvironments();
-  const hasLegacyExperimentRules = features.some(
-    (f) =>
-      getMatchingRules(
-        f,
-        (r) => r.type === "experiment",
-        environments.map((e) => e.id)
-      ).length > 0
-  );
-  const hasNewExperimentRules = features.some(
-    (f) =>
-      getMatchingRules(
-        f,
-        (r) => r.type === "experiment-ref",
-        environments.map((e) => e.id)
-      ).length > 0
-  );
-
-  const defaultValues = {
-    ...defaultRuleValues,
-    ...rule,
-  };
-
-  const [scheduleToggleEnabled, setScheduleToggleEnabled] = useState(
-    (defaultValues.scheduleRules || []).some(
-      (scheduleRule) => scheduleRule.timestamp !== null
-    )
-  );
-
-  const [newRuleOverviewPage, setNewRuleOverviewPage] = useState<boolean>(
-    !defaultType
-  );
-  const [overviewRuleType, setOverviewRuleType] = useState<
-    "force" | "rollout" | "experiment" | "bandit" | ""
-  >("");
-
-  const form = useForm<FeatureRule | NewExperimentRefRule>({
-    defaultValues,
-  });
-  const { apiCall } = useAuth();
-
-  const hasStickyBucketFeature = hasCommercialFeature("sticky-bucketing");
-  const hasMultiArmedBanditFeature = hasCommercialFeature(
-    "multi-armed-bandits"
-  );
-  const usingStickyBucketing = !!settings.useStickyBucketing;
-
-  const type = form.watch("type");
-
-  const hasHashAttributes =
-    attributeSchema.filter((x) => x.hashAttribute).length > 0;
-
-  const experimentId = form.watch("experimentId");
-  const selectedExperiment = experimentsMap.get(experimentId) || null;
-
-  const { data: sdkConnectionsData } = useSDKConnections();
-  const hasSDKWithNoBucketingV2 = !allConnectionsSupportBucketingV2(
-    sdkConnectionsData?.connections,
-    feature.project
-  );
-
-  const prerequisites = form.watch("prerequisites") || [];
-  const [isCyclic, cyclicFeatureId] = useMemo(() => {
-    if (!prerequisites.length) return [false, null];
-    const newFeature = cloneDeep(feature);
-    const revision = revisions?.find((r) => r.version === version);
-    const newRevision = cloneDeep(revision);
-    if (newRevision) {
-      // merge form values into revision
-      const newRule = form.getValues() as FeatureRule;
-      newRevision.rules[environment] = newRevision.rules[environment] || [];
-      newRevision.rules[environment][i] = newRule;
-    }
-    const featuresMap = new Map(features.map((f) => [f.id, f]));
-    return isFeatureCyclic(newFeature, featuresMap, newRevision, [environment]);
-  }, [
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    JSON.stringify(prerequisites),
-    prerequisites.length,
-    features,
-    feature,
-    revisions,
-    version,
-    environment,
-    form,
-    i,
-  ]);
-
-  const [
-    prerequisiteTargetingSdkIssues,
-    setPrerequisiteTargetingSdkIssues,
-  ] = useState(false);
-  const canSubmit = useMemo(() => {
-    return !isCyclic && !prerequisiteTargetingSdkIssues;
-  }, [isCyclic, prerequisiteTargetingSdkIssues]);
-
-  if (showUpgradeModal) {
-    return (
-      <UpgradeModal
-        close={() => setShowUpgradeModal(false)}
-        reason="To enable feature flag scheduling,"
-        source="schedule-feature-flag"
-      />
-    );
-  }
-
-  const ruleTypeOptions = [
-    { label: "Forced Value", value: "force" },
-    { label: "Percentage Rollout", value: "rollout" },
-    { label: "New A/B Experiment", value: "experiment-ref-new" },
-    { label: "Existing A/B Experiment", value: "experiment-ref" },
-  ];
-
-  if (type === "experiment") {
-    ruleTypeOptions.push({
-      label: "A/B Experiment",
-      value: "experiment",
-    });
-  }
-
-  const experimentOptions = experiments
-    .filter(
-      (e) =>
-        e.id === experimentId ||
-        (!e.archived &&
-          e.status !== "stopped" &&
-          (e.project || "") === (feature.project || ""))
-    )
-    .sort((a, b) => b.dateCreated.localeCompare(a.dateCreated))
-    .map((e) => ({
-      label: e.name,
-      value: e.id,
-    }));
-
-  function changeRuleType(v: string) {
-    const existingCondition = form.watch("condition");
-    const existingSavedGroups = form.watch("savedGroups");
-    const newVal = {
-      ...getDefaultRuleValue({
-        defaultValue: getFeatureDefaultValue(feature),
-        ruleType: v,
-        attributeSchema,
-      }),
-      description: form.watch("description"),
-    };
-    if (existingCondition && existingCondition !== "{}") {
-      newVal.condition = existingCondition;
-    }
-    if (existingSavedGroups) {
-      newVal.savedGroups = existingSavedGroups;
-    }
-    form.reset(newVal);
-  }
-
-  const showNewExperimentRuleMessage =
-    hasLegacyExperimentRules &&
-    !hasNewExperimentRules &&
-    (type === "experiment-ref" || type === "experiment-ref-new");
-
-  const canEditTargeting =
-    !!selectedExperiment &&
-    selectedExperiment.linkedFeatures?.length === 1 &&
-    selectedExperiment.linkedFeatures[0] === feature.id &&
-    !selectedExperiment.hasVisualChangesets;
-
-  if (showTargetingModal && canEditTargeting) {
-    const safeToEdit =
-      selectedExperiment.status !== "running" ||
-      !includeExperimentInPayload(selectedExperiment, [feature]);
-
-    return (
-      <EditTargetingModal
-        close={() => setShowTargetingModal(false)}
-        mutate={() => {
-          mutateExperiments();
-          mutate();
-        }}
-        experiment={selectedExperiment}
-        safeToEdit={safeToEdit}
-      />
-    );
-  }
-
-  return (
-    <Modal
-      trackingEventModalType=""
-      open={true}
-      close={close}
-      size="lg"
-      cta="Save"
-      ctaEnabled={canSubmit}
-      bodyClassName="px-4"
-      header={`${rule ? "Edit Rule" : "New Rule"} in ${environment}`}
-      subHeader="You will have a chance to review new rules as a draft before publishing changes."
-      submit={form.handleSubmit(async (values) => {
-        const ruleAction = i === rules.length ? "add" : "edit";
-
-        // If the user built a schedule, but disabled the toggle, we ignore the schedule
-        if (!scheduleToggleEnabled) {
-          values.scheduleRules = [];
-        }
-
-        // Loop through each scheduleRule and convert the timestamp to an ISOString()
-        if (values.scheduleRules?.length) {
-          values.scheduleRules?.forEach((scheduleRule: ScheduleRule) => {
-            if (scheduleRule.timestamp === null) {
-              return;
-            }
-            scheduleRule.timestamp = new Date(
-              scheduleRule.timestamp
-            ).toISOString();
-          });
-
-          // We currently only support a start date and end date, and if both are null, set schedule to empty array
-          if (
-            values.scheduleRules[0].timestamp === null &&
-            values.scheduleRules[1].timestamp === null
-          ) {
-            values.scheduleRules = [];
-          }
-        }
-
-        try {
-          if (values.type === "experiment-ref-new") {
-            // Apply same validation as we do for legacy experiment rules
-            const newRule = validateFeatureRule(
-              {
-                ...values,
-                type: "experiment",
-              },
-              feature
-            );
-            if (newRule) {
-              form.reset({
-                ...newRule,
-                type: "experiment-ref-new",
-                name: values.name,
-              });
-              throw new Error(
-                "We fixed some errors in the rule. If it looks correct, submit again."
-              );
-            }
-
-            // If we're scheduling this rule, always auto start the experiment so it's not stuck in a 'draft' state
-            if (!values.autoStart && values.scheduleRules?.length) {
-              values.autoStart = true;
-            }
-            // If we're starting the experiment immediately, remove any scheduling rules
-            // When we hide the schedule UI the form values don't update, so this resets it if you get into a weird state
-            else if (values.autoStart && values.scheduleRules?.length) {
-              values.scheduleRules = [];
-            }
-
-            // All looks good, create experiment
-            const exp: Partial<ExperimentInterfaceStringDates> = {
-              archived: false,
-              autoSnapshots: true,
-              ...getNewExperimentDatasourceDefaults(
-                datasources,
-                settings,
-                feature.project || ""
-              ),
-              hashAttribute: values.hashAttribute,
-              fallbackAttribute: values.fallbackAttribute || "",
-              goalMetrics: [],
-              secondaryMetrics: [],
-              guardrailMetrics: [],
-              activationMetric: "",
-              name: values.name,
-              hashVersion: hasSDKWithNoBucketingV2 ? 1 : 2,
-              owner: "",
-              status:
-                values.experimentType === "multi-armed-bandit"
-                  ? "draft"
-                  : values.autoStart
-                  ? "running"
-                  : "draft",
-              tags: feature.tags || [],
-              trackingKey: values.trackingKey || feature.id,
-              description: values.description,
-              hypothesis: "",
-              linkedFeatures: [feature.id],
-              attributionModel: settings?.attributionModel || "firstExposure",
-              targetURLRegex: "",
-              ideaSource: "",
-              project: feature.project,
-              variations: values.values.map((v, i) => ({
-                id: uniqId("var_"),
-                key: i + "",
-                name: v.name || (i ? `Variation ${i}` : "Control"),
-                screenshots: [],
-              })),
-              phases: [
-                {
-                  condition: values.condition || "",
-                  savedGroups: values.savedGroups || [],
-                  prerequisites: values.prerequisites || [],
-                  coverage: values.coverage ?? 1,
-                  dateStarted: new Date().toISOString().substr(0, 16),
-                  name: "Main",
-                  namespace: values.namespace || {
-                    enabled: false,
-                    name: "",
-                    range: [0, 1],
-                  },
-                  reason: "",
-                  variationWeights: values.values.map((v) => v.weight),
-                },
-              ],
-              type: values.experimentType,
-            };
-            const res = await apiCall<
-              | { experiment: ExperimentInterfaceStringDates }
-              | { duplicateTrackingKey: true; existingId: string }
-            >(
-              `/experiments${
-                allowDuplicateTrackingKey
-                  ? "?allowDuplicateTrackingKey=true"
-                  : ""
-              }`,
-              {
-                method: "POST",
-                body: JSON.stringify(exp),
-              }
-            );
-
-            if ("duplicateTrackingKey" in res) {
-              setAllowDuplicateTrackingKey(true);
-              throw new Error(
-                "Warning: An experiment with that tracking key already exists. To continue anyway, click 'Save' again."
-              );
-            }
-
-            track("Create Experiment", {
-              source: "experiment-ref-new-rule-modal",
-              numTags: feature.tags?.length || 0,
-              numMetrics: 0,
-              numVariations: values.values.length || 0,
-            });
-
-            // Experiment created, treat it as an experiment ref rule now
-            values = {
-              type: "experiment-ref",
-              description: "",
-              experimentId: res.experiment.id,
-              id: values.id,
-              condition: "",
-              savedGroups: [],
-              enabled: values.enabled ?? true,
-              variations: values.values.map((v, i) => ({
-                value: v.value,
-                variationId: res.experiment.variations[i]?.id || "",
-              })),
-              scheduleRules: values.scheduleRules || [],
-            };
-            mutateExperiments();
-          } else if (values.type === "experiment-ref") {
-            // Validate a proper experiment was chosen and it has a value for every variation id
-            const experimentId = values.experimentId;
-            const exp = experimentsMap.get(experimentId);
-            if (!exp) throw new Error("Must select an experiment");
-
-            const valuesByIndex = values.variations.map((v) => v.value);
-            const valuesByVariationId = new Map(
-              values.variations.map((v) => [v.variationId, v.value])
-            );
-
-            values.variations = exp.variations.map((v, i) => {
-              return {
-                variationId: v.id,
-                value: valuesByVariationId.get(v.id) ?? valuesByIndex[i] ?? "",
-              };
-            });
-
-            delete (values as FeatureRule).condition;
-            delete (values as FeatureRule).savedGroups;
-            delete (values as FeatureRule).prerequisites;
-            // eslint-disable-next-line
-            delete (values as any).value;
-          }
-
-          if (
-            values.scheduleRules &&
-            values.scheduleRules.length === 0 &&
-            !rule?.scheduleRules
-          ) {
-            delete values.scheduleRules;
-          }
-
-          const correctedRule = validateFeatureRule(values, feature);
-          if (correctedRule) {
-            form.reset(correctedRule);
-            throw new Error(
-              "We fixed some errors in the rule. If it looks correct, submit again."
-            );
-          }
-
-          track("Save Feature Rule", {
-            source: ruleAction,
-            ruleIndex: i,
-            environment,
-            type: values.type,
-            hasCondition: values.condition && values.condition.length > 2,
-            hasSavedGroups: !!values.savedGroups?.length,
-            hasPrerequisites: !!values.prerequisites?.length,
-            hasDescription: values.description.length > 0,
-          });
-
-          const res = await apiCall<{ version: number }>(
-            `/feature/${feature.id}/${version}/rule`,
-            {
-              method: i === rules.length ? "POST" : "PUT",
-              body: JSON.stringify({
-                rule: values,
-                environment,
-                i,
-              }),
-            }
-          );
-          await mutate();
-          res.version && setVersion(res.version);
-        } catch (e) {
-          track("Feature Rule Error", {
-            source: ruleAction,
-            ruleIndex: i,
-            environment,
-            type: values.type,
-            hasCondition: values.condition && values.condition.length > 2,
-            hasSavedGroups: !!values.savedGroups?.length,
-            hasPrerequisites: !!values.prerequisites?.length,
-            hasDescription: values.description.length > 0,
-            error: e.message,
-          });
-
-          forceConditionRender();
-
-          throw e;
-        }
-      })}
-    >
-      {newRuleOverviewPage ? (
-        <>
-          <div className="bg-highlight rounded p-3 mb-3">
-            <h5>Select implementation</h5>
-            <RadioCards
-              mt="4"
-              width="100%"
-              options={[
-                {
-                  value: "force",
-                  label: "Force value",
-                  description:
-                    "Target groups of users and give them all the same value",
-                },
-                {
-                  value: "rollout",
-                  label: "Percentage rollout",
-                  description:
-                    "Release to small percent of users while monitoring logs",
-                },
-                {
-                  value: "experiment",
-                  label: "Add Experiment",
-                  description:
-                    "Measure the impact of this feature on your key metrics",
-                },
-                ...(growthbook.isOn("bandits")
-                  ? [
-                      {
-                        value: "bandit",
-                        disabled: !hasMultiArmedBanditFeature,
-                        label: (
-                          <PremiumTooltip
-                            commercialFeature="multi-armed-bandits"
-                            usePortal={true}
-                          >
-                            Add Bandit
-                          </PremiumTooltip>
-                        ),
-                        description:
-                          "Find a winner among many variations on one goal metric",
-                      },
-                    ]
-                  : []),
-              ]}
-              value={overviewRuleType}
-              setValue={(v: "force" | "rollout" | "experiment" | "bandit") => {
-                setOverviewRuleType(v);
-                if (v === "force") {
-                  form.setValue("type", "force");
-                } else if (v === "rollout") {
-                  form.setValue("type", "rollout");
-                } else {
-                  form.setValue("type", undefined);
-                }
-              }}
-            />
-          </div>
-          {overviewRuleType === "experiment" && (
-            <>
-              <h5>Add Experiment</h5>
-              <RadioGroup
-                options={[
-                  {
-                    value: "experiment-ref",
-                    label: "Add existing Experiment",
-                  },
-                  {
-                    value: "experiment-ref-new",
-                    label: "Create new Experiment",
-                  },
-                ]}
-                value={type}
-                setValue={(v) => form.setValue("type", v)}
-              />
-              <Field type="hidden" name="experimentType" value="standard" />
-            </>
-          )}
-          {overviewRuleType === "bandit" && (
-            <>
-              <h5>Add Bandit</h5>
-              <RadioGroup
-                options={[
-                  {
-                    value: "experiment-ref",
-                    label: "Add existing Bandit",
-                  },
-                  {
-                    value: "experiment-ref-new",
-                    label: "Create new Bandit",
-                  },
-                ]}
-                value={type}
-                setValue={(v) => form.setValue("type", v)}
-              />
-              <Field type="hidden" name="experimentType" value="bandit" />
-            </>
-          )}
-        </>
-      ) : (
-        <>
-          <div className="form-group mt-3">
-            <label>Rule Type</label>
-            {!rules[i] ? (
-              <SelectField
-                readOnly={!!rules[i]}
-                value={type}
-                sort={false}
-                onChange={(v) => {
-                  changeRuleType(v);
-                }}
-                options={ruleTypeOptions}
-              />
-            ) : (
-              <div className="border rounded py-2 px-3">
-                {ruleTypeOptions.find((r) => r.value === type)?.label || type}
-                <Field type={"hidden"} {...form.register("type")} />
-              </div>
-            )}
-          </div>
-
-          {type === "experiment-ref" && (
-            <div>
-              {experimentOptions.length > 0 ? (
-                <SelectField
-                  label="Experiment"
-                  initialOption="Choose One..."
-                  options={experimentOptions}
-                  readOnly={!!rules[i]}
-                  disabled={!!rules[i]}
-                  required
-                  sort={false}
-                  value={experimentId || ""}
-                  onChange={(experimentId) => {
-                    const exp = experimentsMap.get(experimentId);
-                    if (exp) {
-                      const controlValue = getFeatureDefaultValue(feature);
-                      const variationValue = getDefaultVariationValue(
-                        controlValue
-                      );
-                      form.setValue("experimentId", experimentId);
-                      form.setValue(
-                        "variations",
-                        exp.variations.map((v, i) => ({
-                          variationId: v.id,
-                          value: i ? variationValue : controlValue,
-                        }))
-                      );
-                    }
-                  }}
-                  formatOptionLabel={({ value, label }) => {
-                    const exp = experimentsMap.get(value);
-                    if (exp) {
-                      return (
-                        <div className="d-flex flex-wrap">
-                          <div className="flex">
-                            <strong>{exp.name}</strong>
-                          </div>
-                          <div className="ml-4 text-muted">
-                            Created: {date(exp.dateCreated)}
-                          </div>
-                          <div className="ml-auto">
-                            <StatusIndicator
-                              archived={exp.archived}
-                              status={exp.status}
-                            />
-                          </div>
-                        </div>
-                      );
-                    }
-                    return label;
-                  }}
-                />
-              ) : !rules[i] ? (
-                <div className="alert alert-warning">
-                  <div className="d-flex align-items-center">
-                    {experiments.length > 0
-                      ? `You don't have any elegible experiments yet.`
-                      : `You don't have any existing experiments yet.`}{" "}
-                    <button
-                      type="button"
-                      className="btn btn-primary ml-auto"
-                      onClick={(e) => {
-                        e.preventDefault();
-                        changeRuleType("experiment-ref-new");
-                      }}
-                    >
-                      Create New Experiment
-                    </button>
-                  </div>
-                </div>
-              ) : (
-                <div className="alert alert-danger">
-                  Could not find this experiment. Has it been deleted?
-                </div>
-              )}
-
-              {selectedExperiment && rules[i] && (
-                <div className="appbox px-3 pt-3 bg-light">
-                  {!canEditTargeting && (
-                    <div className="alert alert-info">
-                      <Link
-                        href={`/experiment/${selectedExperiment.id}#overview`}
-                        className="alert-link"
-                      >
-                        View the Experiment
-                        <FaExternalLinkAlt />
-                      </Link>{" "}
-                      to make changes to assignment or targeting conditions.
-                    </div>
-                  )}
-                  <TargetingInfo
-                    experiment={selectedExperiment}
-                    editTargeting={
-                      canEditTargeting
-                        ? () => {
-                            setShowTargetingModal(true);
-                          }
-                        : null
-                    }
-                  />
-                </div>
-              )}
-              {selectedExperiment && (
-                <div className="form-group">
-                  <label>Variation Values</label>
-                  <div className="mb-3 bg-light border p-3">
-                    {selectedExperiment.variations.map((v, i) => (
-                      <FeatureValueField
-                        key={v.id}
-                        label={v.name}
-                        id={v.id}
-                        value={form.watch(`variations.${i}.value`) || ""}
-                        setValue={(v) =>
-                          form.setValue(`variations.${i}.value`, v)
-                        }
-                        valueType={feature.valueType}
-                        feature={feature}
-                        renderJSONInline={false}
-                      />
-                    ))}
-                  </div>
-                </div>
-              )}
-            </div>
-          )}
-
-          {type === "experiment-ref-new" && (
-            <>
-              {growthbook.isOn("bandits") && (
-                <div className="bg-highlight rounded py-3 px-3 mb-4">
-                  <ButtonSelectField
-                    buttonType="card"
-                    value={form.watch("experimentType") || ""}
-                    setValue={(v) =>
-                      form.setValue("experimentType", v as ExperimentType)
-                    }
-                    options={[
-                      {
-                        label: (
-                          <div
-                            className="mx-3 d-flex flex-column align-items-center justify-content-center"
-                            style={{ minHeight: 90 }}
-                          >
-                            <div className="h4">
-                              {form.watch("experimentType") === "standard" && (
-                                <FaRegCircleCheck
-                                  size={18}
-                                  className="check text-success mr-2"
-                                />
-                              )}
-                              Experiment
-                            </div>
-                            <div className="small">
-                              Variation weights are constant throughout the
-                              experiment
-                            </div>
-                          </div>
-                        ),
-                        value: "standard",
-                      },
-                      {
-                        label: (
-                          <div
-                            className="mx-3 d-flex flex-column align-items-center justify-content-center"
-                            style={{ minHeight: 90 }}
-                          >
-                            <div className="h4">
-                              <PremiumTooltip
-                                commercialFeature="multi-armed-bandits"
-                                body={
-                                  !usingStickyBucketing &&
-                                  hasStickyBucketFeature ? (
-                                    <div>
-                                      Enable Sticky Bucketing in your
-                                      organization settings to run a Bandit.
-                                    </div>
-                                  ) : null
-                                }
-                                usePortal={true}
-                              >
-                                {form.watch("experimentType") ===
-                                  "multi-armed-bandit" && (
-                                  <FaRegCircleCheck
-                                    size={18}
-                                    className="check text-success mr-2"
-                                  />
-                                )}
-                                Bandit
-                              </PremiumTooltip>
-                            </div>
-
-                            <div className="small">
-                              Variations with better results receive more
-                              traffic during the experiment
-                            </div>
-                          </div>
-                        ),
-                        value: "multi-armed-bandit",
-                        disabled:
-                          !hasMultiArmedBanditFeature || !usingStickyBucketing,
-                      },
-                    ]}
-                  />
-                </div>
-              )}
-
-              <Field
-                label={
-                  form.watch("experimentType") === "multi-armed-bandit"
-                    ? "Bandit Name"
-                    : "Experiment Name"
-                }
-                {...form.register("name")}
-                required
-              />
-            </>
-          )}
-
-          {type !== "experiment-ref" && (
-            <Field
-              label="Description"
-              textarea
-              minRows={1}
-              {...form.register("description")}
-              placeholder="Short human-readable description of the rule"
-            />
-          )}
-          {type === "force" && (
-            <FeatureValueField
-              label="Value to Force"
-              id="value"
-              value={form.watch("value")}
-              setValue={(v) => form.setValue("value", v)}
-              valueType={feature.valueType}
-              feature={feature}
-              renderJSONInline={true}
-            />
-          )}
-
-          {type === "rollout" && (
-            <div>
-              <FeatureValueField
-                label="Value to roll out"
-                id="value"
-                value={form.watch("value")}
-                setValue={(v) => form.setValue("value", v)}
-                valueType={feature.valueType}
-                feature={feature}
-                renderJSONInline={true}
-              />
-              <div className="appbox mt-4 mb-4 px-3 pt-3 bg-light">
-                <RolloutPercentInput
-                  value={form.watch("coverage") || 0}
-                  setValue={(coverage) => {
-                    form.setValue("coverage", coverage);
-                  }}
-                  className="mb-1"
-                />
-                <SelectField
-                  label="Assign value based on attribute"
-                  options={attributeSchema
-                    .filter((s) => !hasHashAttributes || s.hashAttribute)
-                    .map((s) => ({ label: s.property, value: s.property }))}
-                  value={form.watch("hashAttribute")}
-                  onChange={(v) => {
-                    form.setValue("hashAttribute", v);
-                  }}
-                  helpText={
-                    "The globally unique tracking key for the experiment"
-                  }
-                />
-              </div>
-            </div>
-          )}
-
-          {(type !== "experiment-ref" && type !== "experiment-ref-new") ||
-          rule?.scheduleRules?.length ? (
-            <ScheduleInputs
-              defaultValue={defaultValues.scheduleRules || []}
-              onChange={(value) => form.setValue("scheduleRules", value)}
-              scheduleToggleEnabled={scheduleToggleEnabled}
-              setScheduleToggleEnabled={setScheduleToggleEnabled}
-              setShowUpgradeModal={setShowUpgradeModal}
-              title="Add scheduling to automatically enable/disable this rule"
-            />
-          ) : null}
-
-          {(type === "experiment" || type === "experiment-ref-new") && (
-            <>
-              <div className="mt-4 mb-4">
-                <Field
-                  label="Tracking Key"
-                  {...form.register(`trackingKey`)}
-                  placeholder={feature.id}
-                  helpText="Unique identifier for this experiment, used to track impressions and analyze results"
-                />
-                <div className="d-flex" style={{ gap: "2rem" }}>
-                  <SelectField
-                    label="Assign value based on attribute"
-                    containerClassName="flex-1"
-                    options={attributeSchema
-                      .filter((s) => !hasHashAttributes || s.hashAttribute)
-                      .map((s) => ({ label: s.property, value: s.property }))}
-                    value={form.watch("hashAttribute")}
-                    onChange={(v) => {
-                      form.setValue("hashAttribute", v);
-                    }}
-                    helpText={
-                      "Will be hashed together with the Tracking Key to determine which variation to assign"
-                    }
-                  />
-                  <FallbackAttributeSelector
-                    form={form}
-                    attributeSchema={attributeSchema}
-                  />
-                </div>
-              </div>
-              {hasSDKWithNoBucketingV2 && (
-                <HashVersionSelector
-                  value={(form.watch("hashVersion") || 1) as 1 | 2}
-                  onChange={(v) => form.setValue("hashVersion", v)}
-                  project={feature.project}
-                />
-              )}
-              <hr />
-            </>
-          )}
-
-          {!(
-            type === "experiment" ||
-            type === "experiment-ref" ||
-            type === "experiment-ref-new"
-          ) && <hr />}
-
-          {type !== "experiment-ref" && (
-            <div className="mt-4">
-              <SavedGroupTargetingField
-                value={form.watch("savedGroups") || []}
-                setValue={(savedGroups) =>
-                  form.setValue("savedGroups", savedGroups)
-                }
-                project={feature.project || ""}
-              />
-              <hr />
-              <ConditionInput
-                defaultValue={form.watch("condition") || ""}
-                onChange={(value) => form.setValue("condition", value)}
-                key={conditionKey}
-                project={feature.project || ""}
-              />
-              <hr />
-              <PrerequisiteTargetingField
-                value={form.watch("prerequisites") || []}
-                setValue={(prerequisites) =>
-                  form.setValue("prerequisites", prerequisites)
-                }
-                feature={feature}
-                revisions={revisions}
-                version={version}
-                environments={[environment]}
-                setPrerequisiteTargetingSdkIssues={
-                  setPrerequisiteTargetingSdkIssues
-                }
-              />
-              {(type === "experiment" || type === "experiment-ref-new") && (
-                <hr />
-              )}
-            </div>
-          )}
-          {isCyclic && (
-            <div className="alert alert-danger">
-              <FaExclamationTriangle /> A prerequisite (
-              <code>{cyclicFeatureId}</code>) creates a circular dependency.
-              Remove this prerequisite to continue.
-            </div>
-          )}
-
-          {(type === "experiment" || type === "experiment-ref-new") && (
-            <div>
-              {namespaces && namespaces.length > 0 && (
-                <NamespaceSelector
-                  form={form}
-                  trackingKey={form.watch("trackingKey") || feature.id}
-                  featureId={feature.id}
-                  formPrefix=""
-                />
-              )}
-              <div className="mb-4">
-                <FeatureVariationsInput
-                  defaultValue={getFeatureDefaultValue(feature)}
-                  valueType={feature.valueType}
-                  coverage={form.watch("coverage") || 0}
-                  setCoverage={(coverage) =>
-                    form.setValue("coverage", coverage)
-                  }
-                  setWeight={(i, weight) =>
-                    form.setValue(`values.${i}.weight`, weight)
-                  }
-                  variations={
-                    form
-                      .watch("values")
-                      .map((v: ExperimentValue & { id?: string }) => {
-                        return {
-                          value: v.value || "",
-                          name: v.name,
-                          weight: v.weight,
-                          id: v.id || generateVariationId(),
-                        };
-                      }) || []
-                  }
-                  setVariations={(variations) =>
-                    form.setValue("values", variations)
-                  }
-                  feature={feature}
-                  simple={form.watch("experimentType") === "multi-armed-bandit"}
-                />
-              </div>
-            </div>
-          )}
-          {type === "experiment-ref-new" &&
-            form.watch("experimentType") !== "multi-armed-bandit" && (
-              <div className="mb-3">
-                <Toggle
-                  value={form.watch("autoStart")}
-                  setValue={(v) => form.setValue("autoStart", v)}
-                  id="auto-start-new-experiment"
-                />{" "}
-                <label
-                  htmlFor="auto-start-new-experiment"
-                  className="text-dark"
-                >
-                  Start Experiment Immediately
-                </label>
-                <div>
-                  <small className="form-text text-muted">
-                    If On, the experiment will start serving traffic as soon as
-                    the feature is published. Leave Off if you want to make
-                    additional changes before starting.
-                  </small>
-                </div>
-                {!form.watch("autoStart") && (
-                  <div>
-                    <hr />
-                    <ScheduleInputs
-                      defaultValue={defaultValues.scheduleRules || []}
-                      onChange={(value) =>
-                        form.setValue("scheduleRules", value)
-                      }
-                      scheduleToggleEnabled={scheduleToggleEnabled}
-                      setScheduleToggleEnabled={setScheduleToggleEnabled}
-                      setShowUpgradeModal={setShowUpgradeModal}
-                    />
-                  </div>
-                )}
-              </div>
-            )}
-        </>
-      )}
-    </Modal>
-  );
-}
